Data set: The GRDC Catalogue

The GRDC catalogue

data("GRDCcatalogue")

Format

A data frame with 9252 stations (rows) and 46 metadata fields.

grdc_no

GRDC station number

wmo_reg

WMO region

sub_reg

WMO subregion

mix_reg

Combination of row 2+3

nat_id

national station ID

river

river name

station

station name

country_code

country code (ISO 3166)

lat

latitude in degrees

long

longitude in degrees

area

catchment size Km2

altitude

height of gauge zero above sea level m

ds_stat_no

GRDC_No of next downstream GRDC station

w_level

water level data available in addition to discharge

d_start

daily data available from

d_end

daily data available until

d_yrs

# years of daily data

d_miss

percentage of missing values (daily data)

m_start

monthly data available from

m_end

monthly data available until

m_yrs

# years of monthly data

m_miss

percentage of missing values (monthly data)

t_start

totally earliest data available

t_end

totally latest data available

t_yrs

# years (maximum using daily and monthly data)

lta_discharge

mean annual streamflow m3/s

r_volume_yr

mean annual volume km3

r_height_yr

mean annual depth mm

proc_tyrs

percentage of t_yrs used for calculation of lta_discharge, r_vol_yr and r_height_yr (only years with > 9 month i.e. > 75 % of monthly data are considered)

proc_tmon

percentage of # of month in proc_tyrs * t_yrs years used for calculation of lta_discharge, r_vol_yr and r_height_yr (minimum 75 % are considered)

f_import

date of first import

f_im_yr

year of first import

l_import

date of last change

l_im_yr

year of last change

provider_id

provider number (to be resolved with address database)

acsys

1 if station belongs to the subset of the Arctic Runoff Data Base in the WCRP Arctic Climate System Study

flux2ocean

NA

gems

1 if station corresponds to a GEMS/Water water quality station

gcos_gtn.h

1 if station belongs to the subset of the Global Terrestrial Network for River Discharge within GTN-H

statistics

NA

Pristine.Basins

1 if station belongs to the subset Pristine Basins

GRDC.Reference.Dataset

1 if station belongs to the subset GRDC Reference Data set

AdaptAlp

1 if station belongs to the subset AdaptAlp

ArctHycos

1 if station belongs to the subset ArcticHYCOS

EWA

NA

BALTEX

NA

Source

http://www.bafg.de/GRDC/EN/02_srvcs/21_tmsrs/211_ctlgs/catalogues_node.html