Data source: Global Runoff Data Centre catalogue

This function interfaces the Global Runoff Data Centre database which provides river discharge data for almost 1000 sites over 157 countries.

catalogueGRDC()

Value

This function returns a data frame made with the following columns:

grdc_no: GRDC station number

wmo_reg: WMO region

sub_reg: WMO subregion

river: river name

station: station name

country: 2-letter country code (ISO 3166)

lat: latitude, decimal degree

long: longitude, decimal degree

area: catchment size, km2

altitude: height of gauge zero, m above sea level

d_start: daily data available from year

d_end: daily data available until year

d_yrs: length of time series, daily data

d_miss: percentage of missing values (daily data)

m_start: monthly data available from

m_end: monthly data available until

m_yrs: length of time series, monthly data

m_miss: percentage of missing values (monthly data)

t_start: earliest data available

t_end: latest data available

t_yrs: maximum length of time series, daily and monthly data

lta_discharge: mean annual streamflow, m3/s

r_volume_yr: mean annual volume, km3

r_height_yr: mean annual runoff depth, mm

Author

Claudia Vitolo

Examples

if (FALSE) {
  # Retrieve the catalogue
  GRDC_catalogue_all <- catalogueGRDC()
}