Interface for the MOPEX database of Daily Time Series

This function extract the dataset containing daily rainfall and streamflow discharge at one of the MOPEX locations.

tsMOPEX(id, MAP = TRUE)

Arguments

id

String for the station ID number (USGS_ID)

MAP

Boolean, TRUE by default. If FALSE it looks for data through all the 1861 potential MOPEX basins. If TRUE, it looks for data through the 438 MOPEX basins with MAP estimates.

Value

If MAP = FALSE, this function returns a time series of daily streamflow discharge (Q, in mm). If MAP = TRUE, this function returns a data frame containing the following columns (as zoo object):

Date

Format is "yyyymmdd"

P

Mean areal precipitation (mm)

E

Climatic potential evaporation (mm, based NOAA Freewater Evaporation Atlas)

Q

Daily streamflow discharge (mm)

T_max

Daily maximum air temperature (Celsius)

T_min

Daily minimum air temperature (Celsius)

Author

Claudia Vitolo

Examples

if (FALSE) { # \dontrun{
  BroadRiver <- tsMOPEX(id = "01048000")
} # }