Interface for the MOPEX database of Daily Time Series

This function extract the dataset containing daily rainfall and streamflow discharge at one of the MOPEX locations.

tsSEPA(stationID, plotOption = FALSE, timeExtent = NULL)

Arguments

stationID
hydrometric reference number (string)

plotOption
boolean to define whether to plot the results. By default this is set to TRUE.

timeExtent
is a vector of dates and times for which the data should be retrieved

Value

The function returns river level data in a zoo object.

Examples

# NOT RUN {
  sampleID <- catalogueSEPA()$stationId[1]
  sampleTS <- tsSEPA(stationID = sampleID, plotOption = TRUE)
# }