Data set: The GRDC Catalogue

The GRDC catalogue

data("GRDCcatalogue")

Format

A data frame with 9252 stations (rows) and 46 metadata fields.

grdc_no: GRDC station number wmo_reg: WMO region sub_reg: WMO subregion nat_id: national station ID river: river name station: station name country_code: country code (ISO 3166) lat: latitude, decimal degree long: longitude, decimal degree area: catchment size, km2 altitude: height of gauge zero, m above sea level ds_stat_no: GRDC station number of next downstream GRDC station d_start: daily data available from year d_end: daily data available until year d_yrs: length of time series, daily data d_miss: percentage of missing values (daily data) m_start: monthly data available from m_end: monthly data available until m_yrs: length of time series, monthly data m_miss: percentage of missing values (monthly data) t_start: earliest data available t_end: latest data available t_yrs: maximum length of time series, daily and monthly data lta_discharge: mean annual streamflow, m3/s r_volume_yr: mean annual volume, km3 r_height_yr: mean annual runoff depth, mm

Source

http://www.bafg.de/GRDC/EN/02_srvcs/21_tmsrs/211_ctlgs/catalogues_node.html