Data source: Global Runoff Data Centre catalogue

This function interfaces the Global Runoff Data Centre database which provides river discharge data for about 9000 sites over 157 countries.

catalogueGRDC(areaBox = NULL, columnName = NULL, columnValue = NULL,
  useCachedData = TRUE)

Arguments

areaBox
bounding box, a list made of 4 elements: minimum longitude (lonMin), minimum latitude (latMin), maximum longitude (lonMax), maximum latitude (latMax)

columnName
name of the column to filter

columnValue
value to look for in the column named columnName

useCachedData
logical, set to TRUE to use cached data, set to FALSE to retrive data from online source. This is TRUE by default.

Value

This function returns a data frame made of 9481 rows (gauging stations, as per October 2017) and 26 columns:

grdc_no: GRDC station number

wmo_reg: WMO region

sub_reg: WMO subregion

nat_id: national station ID

river: river name

station: station name

country_code: country code (ISO 3166)

lat: latitude, decimal degree

long: longitude, decimal degree

area: catchment size, km2

altitude: height of gauge zero, m above sea level

ds_stat_no: GRDC station number of next downstream GRDC station

d_start: daily data available from year

d_end: daily data available until year

d_yrs: length of time series, daily data

d_miss: percentage of missing values (daily data)

m_start: monthly data available from

m_end: monthly data available until

m_yrs: length of time series, monthly data

m_miss: percentage of missing values (monthly data)

t_start: earliest data available

t_end: latest data available

t_yrs: maximum length of time series, daily and monthly data

lta_discharge: mean annual streamflow, m3/s

r_volume_yr: mean annual volume, km3

r_height_yr: mean annual runoff depth, mm

Examples

# NOT RUN {
  # Retrieve the whole catalogue
  GRDC_catalogue_all <- catalogueGRDC()

  # Define a bounding box
  areaBox <- raster::extent(-3.82, -3.63, 52.41, 52.52)
  # Filter the catalogue based on bounding box
  GRDC_catalogue_bbox <- catalogueGRDC(areaBox = areaBox)

  # Get only catchments with area above 5000 Km2
  GRDC_catalogue_area <- catalogueGRDC(columnName = "area",
                                       columnValue = ">= 5000")

  # Get only catchments within river Thames
  GRDC_catalogue_river <- catalogueGRDC(columnName = "river",
                                        columnValue = "Thames")
# }